+++ service/controller/rolebindingtemplate/resource/rolebinding/error.go
@@ -4,6 +4,15 @@ import (
 	"github.com/giantswarm/microerror"
 )
 
+var roleBindingCreationFailedError = &microerror.Error{
+	Kind: "roleBindingCreationFailedError",
+}
+
+// IsRoleBindingCreationFailed asserts roleBindingCreationFailedError.
+func IsRoleBindingCreationFailed(err error) bool {
+	return microerror.Cause(err) == roleBindingCreationFailedError
+}
+
 var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
 }